+++ Assets/Scripts/Enemy/Zoomer.cs
@@ -11,7 +11,7 @@ public class Zoomer : MonoBehaviour
 
     public int damage = 1;
 
-    public int health = 1;
+    public float health = 1;
 
     // Start is called before the first frame update
     void Start()
